fix(activities): archiving a record must not strand the files on it
The end-of-work review found that the upload fix landed on a SHARED helper.
ensureAttachmentParentWritable is the write arm of resolveAttachmentParent,
so making it live also bound every non-read caller: ArchiveAttachment,
UpdateAttachmentMetadata, and the extraction read's claim and finish.

Two things that would have broken. A misfiled document on a deal that is
later archived could never be removed — there is no unarchive verb, and
erasure is person-scoped, so the file would sit there for good. And a
document reading in flight when its parent was archived would fail its
finish, fail its retry's claim with ErrNotFound rather than ErrConflict so
the reclaim arm never runs, and leave the row `running` behind the
in-flight unique index for ever.

That is the same grant-versus-revoke error the four reverted gates had, in
the one site that stayed. The gates are now two: the upload takes
ensureAttachmentParentWritableLive because it ADDS, and everything that
removes or amends something already filed keeps the permissive one.

Also from the review:

  - fillDiscoveredFields carries the subject's liveness in the INSERT, not
    only in the probe the entry point ran. person_profile_field is declared
    PII and erasure DELETES its rows while stamping the person archived, so
    a discovery approved before that commit and applied after it put the
    erased subject's details straight back — in the window between two
    statements rather than over the hours the entry gate closes.
  - the new compose suite asserted a premise that was false in both halves.
    It named resolveVisibleAttachmentParent, which does not exist (the
    second invented name on this branch), and claimed the file could never
    be read back — the read arm uses the NON-live probe, so it always
    could. It also claimed only unbounded seats were exposed; an owner
    passes the scope clause too. Rewritten to say what it holds, and the
    read-side question is named as open rather than asserted as answered.
  - the size_band arm of "nothing landed" was vacuous: the proposal carried
    no facts, so the column could not have moved however the gate behaved.
    It now carries an employee_range fact against a real site_read, and the
    assertion fails against the old code like the other three.
  - both suites retire records through ArchiveOrganization/ArchivePerson/
    ArchiveDeal rather than stamping archived_at themselves.
  - archivedOrgColumn folded back into orgColumn, and asArchiver reused
    from researchclaim rather than declared a second time.

Refs #2201, #2145
